fix: 전체 viewer에 링크 기능 절대경로로 수정
