+++ src/page/qna-detail/components/Markdown/MdViewer.tsx
@@ -5,6 +5,7 @@ import "prismjs/themes/prism.css"
 import "@toast-ui/editor-plugin-code-syntax-highlight/dist/toastui-editor-plugin-code-syntax-highlight.css"
 import "@toast-ui/editor-plugin-code-syntax-highlight/dist/toastui-editor-plugin-code-syntax-highlight-all.js"
 import codeSyntaxHighlight from "@toast-ui/editor-plugin-code-syntax-highlight/dist/toastui-editor-plugin-code-syntax-highlight-all.js"
+import { handleViewerLink } from "@/util/editor"
 
 export interface ViewerProps {
   content: string
@@ -16,29 +17,7 @@ const MdViewer: React.FC<ViewerProps> = ({ content }) => {
       {content && (
         <div
           className="[&_.toastui-editor-contents]:text-[20px]"
-          onClick={(e) => {
-            if ((e.target as HTMLElement).tagName !== "A") return
-
-            const target = e.target as HTMLAnchorElement
-
-            if (
-              !target.href.startsWith("http://") ||
-              !target.href.startsWith("https://")
-            ) {
-              e.preventDefault()
-
-              const origin = `${window.location.origin}/question`
-
-              const targetURL = target.href.replace(`${origin}/`, "")
-              const newLink = `https://${targetURL}`
-
-              const linkElement = document.createElement("a")
-              linkElement.href = newLink
-              linkElement.target = "_blank"
-
-              linkElement.click()
-            }
-          }}
+          onClick={(e) => handleViewerLink("question")(e)}
         >
           <Viewer
             initialValue={content || " "}
