+++ src/page/user-profile/components/introduction/MdViewer.tsx
@@ -5,6 +5,7 @@ import "prismjs/themes/prism.css"
 import "@toast-ui/editor-plugin-code-syntax-highlight/dist/toastui-editor-plugin-code-syntax-highlight.css"
 import "@toast-ui/editor-plugin-code-syntax-highlight/dist/toastui-editor-plugin-code-syntax-highlight-all.js"
 import codeSyntaxHighlight from "@toast-ui/editor-plugin-code-syntax-highlight/dist/toastui-editor-plugin-code-syntax-highlight-all.js"
+import { handleViewerLink } from "@/util/editor"
 
 export interface ViewerProps {
   content: string
@@ -14,7 +15,10 @@ const MdViewer: React.FC<ViewerProps> = ({ content }) => {
   return (
     <div>
       {content && (
-        <div className="[&_.toastui-editor-contents]:text-[16px]">
+        <div
+          className="[&_.toastui-editor-contents]:text-[16px]"
+          onClick={(e) => handleViewerLink("profile")(e)}
+        >
           <Viewer
             initialValue={content || " "}
             /*@ts-ignore*/
